+++ lib/declare_schema.rb
@@ -162,11 +162,7 @@ module DeclareSchema
     end
 
     def current_adapter(model_class = ActiveRecord::Base)
-      if Rails::VERSION::MAJOR >= 6.1
-        model_class.connection_db_config.adapter
-      else
-        model_class.connection_config[:adapter]
-      end
+      model_class.connection_db_config.adapter
     end
   end
 end